IMPORTANT TPS CHANGES IN THIS VERSION
In previous versions,  the TPS numbers were based on writes and reads collected from cfstats as total reads and writes on the specific node.   This is not the case.  Reads and writes in the cfstats file gives the number of read and write requests to the node  as a coordinating node.  This impacts the TPS numbers significantly, especially clusters with multiple DCs.